+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [9.1.0] IN PROGRESS
 
+* Fetch parent and child sub instances in one query. Fixes UIIN-1902.
+
 ## [9.0.0](https://github.com/folio-org/ui-inventory/tree/v9.0.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v8.0.0...v9.0.0)